MinoControllerのミノ生成部分をMinoGeneraterとして分割。
