+++ Assets/Scripts/Mino/MinoGenerater.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 00396e00ee0196a4ebd298998bc9738f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
